Migrating transformation from DIM->DTR out to a separate class.
