+++ src/scripts/destinyTrackerApi/gunTransformer.js
@@ -1,0 +1,35 @@
+class gunTransformer {
+  translateToDtrGun(gun) {
+    return {
+      referenceId: gun.hash,
+      roll: this.getDtrRoll(gun)
+    };
+  }
+
+  getRollAndPerks(gun) {
+    return {
+      roll: this.getDtrRoll(gun),
+      selectedPerks: this.getDtrPerks(gun),
+      referenceId: gun.hash,
+      instanceId: gun.id,
+    };
+  }
+
+  getDtrPerks(gun) {
+    if (!gun.talentGrid) {
+      return null;
+    }
+
+    return gun.talentGrid.dtrPerks;
+  }
+
+  getDtrRoll(gun) {
+    if (!gun.talentGrid) {
+      return null;
+    }
+
+    return gun.talentGrid.dtrRoll;
+  }
+}
+
+export { gunTransformer };
