+++ src/scripts/services/dimDestinyTrackerService.factory.js
@@ -1,42 +1,9 @@
 import angular from 'angular';
 import _ from 'underscore';
+import { gunTransformer } from '../destinyTrackerApi/gunTransformer.js';
 
 angular.module('dimApp')
   .factory('dimDestinyTrackerService', DestinyTrackerService);
-
-class gunTransformer {
-  translateToDtrGun(gun) {
-    return {
-      referenceId: gun.hash,
-      roll: this.getDtrRoll(gun)
-    };
-  }
-
-  getRollAndPerks(gun) {
-    return {
-      roll: this.getDtrRoll(gun),
-      selectedPerks: this.getDtrPerks(gun),
-      referenceId: gun.hash,
-      instanceId: gun.id,
-    };
-  }
-
-  getDtrPerks(gun) {
-    if (!gun.talentGrid) {
-      return null;
-    }
-
-    return gun.talentGrid.dtrPerks;
-  }
-
-  getDtrRoll(gun) {
-    if (!gun.talentGrid) {
-      return null;
-    }
-
-    return gun.talentGrid.dtrRoll;
-  }
-}
 
 class reviewDataCache {
   constructor() {
